requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anykeystore==0.2
apex @ file:///home/lilipan/ling/SHAP/reference/apex-22.04-dev
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
av==10.0.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
blis==0.7.9
Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175564434/work
brotlipy==0.7.0
cachetools==5.3.0
captum==0.5.0
catalogue==2.0.8
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
chardet @ file:///tmp/build/80754af9/chardet_1607706768982/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
cloudpickle==2.2.1
colorama==0.4.6
confection==0.0.4
cryptacular==1.6.2
cryptography @ file:///croot/cryptography_1677533068310/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
cymem==2.0.7
Cython @ file:///croot/cython_1676568029361/work
debugpy @ file:///tmp/build/80754af9/debugpy_1637091426235/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml==0.7.1
docker-pycreds==0.4.0
docopt==0.6.2
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649908585034/work
fastai==2.5.6
fastcore==1.4.5
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.12.2
flatbuffers==23.3.3
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.25.0
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.1.1
future==0.18.2
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.3
google-auth-oauthlib==0.4.6
greenlet==3.0.3
grpcio==1.54.0
h5py @ file:///tmp/abs_4aewd3wzey/croots/recipe/h5py_1659091371897/work
huggingface-hub==0.15.1
hupper==1.12
hydra-core==1.2.0
idna @ file:///croot/idna_1666125576474/work
imageio==2.31.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipdb==0.13.13
ipykernel @ file:///opt/conda/conda-bld/ipykernel_1662361798230/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1651240553635/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
jedi @ file:///tmp/build/80754af9/jedi_1644299024593/work
Jinja2==3.1.2
joblib==1.2.0
jsonpickle==1.5.2
jupyter_client @ file:///croot/jupyter_client_1676329080601/work
jupyter_core @ file:///croot/jupyter_core_1668084443243/work
keras==2.11.0
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
langcodes==3.3.0
libclang==16.0.0
lightly-utils==0.0.2
llvmlite==0.39.1
Markdown==3.4.3
MarkupSafe==2.1.2
-e git+https://github.com/KaihuaTang/Scene-Graph-Benchmark.pytorch.git@4b6b71a90d4198d9dae574d42b062a5e534da291#egg=maskrcnn_benchmark
matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1647441664166/work
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
multidict==6.0.4
munch==2.5.0
munkres==1.1.4
murmurhash==1.0.9
nest-asyncio @ file:///croot/nest-asyncio_1672387112409/work
networkx==2.6.3
numba==0.56.4
numexpr @ file:///croot/numexpr_1668713893690/work
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1653915516269/work
oauthlib==3.2.2
omegaconf==2.2.3
onedrivedownloader==1.1.3
opencv-python==4.6.0.66
packaging @ file:///croot/packaging_1671697413597/work
pandas==1.3.5
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
PasteDeploy==3.1.0
pathtools==0.1.2
pathy==0.10.2
pbkdf2==1.3
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.2.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
preshed==3.0.8
prettytable==3.7.0
promise==2.3
prompt-toolkit @ file:///croot/prompt-toolkit_1672387306916/work
protobuf==3.20.3
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.9
pyDeprecate==0.3.1
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
pyramid==2.0.2
pyramid-mailer==0.15.1
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python3-openid==3.2.0
pytz==2022.4
PyWavelets==1.3.0
PyYAML @ file:///croot/pyyaml_1670514731622/work
pyzmq @ file:///opt/conda/conda-bld/pyzmq_1657724186960/work
quadprog==0.1.11
regex==2023.6.3
repoze.sendmail==4.4.1
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.1
rsa==4.9
sacred==0.8.2
sacremoses==0.0.53
safetensors==0.4.1
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.0
sentry-sdk==1.26.0
setproctitle==1.3.2
shap==0.41.0
shortuuid==1.0.11
six @ file:///tmp/build/80754af9/six_1644875935023/work
slicer==0.0.7
smart-open==6.3.0
smmap==5.0.0
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.25
srsly==2.4.6
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-logger==0.1.0
tensorboard-plugin-wit==1.8.1
thinc==8.1.10
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.6.13
tokenizers==0.10.3
tomli==2.0.1
torch==1.9.1
torch-tb-profiler==0.4.3
torchaudio==0.9.0a0+a85b239
torchmetrics==0.7.3
torchtext==0.4.0
torchvision==0.10.1
tornado @ file:///opt/conda/conda-bld/tornado_1662061693373/work
tqdm @ file:///opt/conda/conda-bld/tqdm_1664392687731/work
traitlets @ file:///croot/traitlets_1671143879854/work
transaction==4.0
transformers==4.3.0
translationstring==1.4
ttach==0.0.3
typer==0.7.0
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
urllib3 @ file:///croot/urllib3_1673575502006/work
velruse==1.1.1
venusian==3.1.0
wandb==0.12.21
wasabi==1.1.2
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
WebOb==1.8.7
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
wtforms-recaptcha==0.3.2
xlwt==1.3.0
yacs @ file:///tmp/build/80754af9/yacs_1634047592950/work
yarl==1.9.2
zipp==3.15.0
zope.deprecation==5.0
zope.interface==6.1
zope.sqlalchemy==3.1
